requirements.txt
numpy==1.24.3
opencv-python-headless==4.8.0.76
PyQt5==5.15.9
torch==2.0.1
torchvision==0.15.2
pandas==2.0.3
matplotlib==3.7.2
scikit-image==0.21.0
pillow==10.0.0
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0 